Added new coretypes project. Binding urn from url route instead of manually finding it in route values. Added NUnit analyzers
