+++ ConcernsCaseWork/Service.Redis.Tests/Service.Redis.Tests.csproj
@@ -13,6 +13,7 @@
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.0" />
         <PackageReference Include="Moq" Version="4.18.2" />
         <PackageReference Include="NUnit" Version="3.13.3" />
+        <PackageReference Include="NUnit.Analyzers" Version="3.3.0" />
         <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
         <PackageReference Include="coverlet.collector" Version="3.1.2">
           <PrivateAssets>all</PrivateAssets>
